+++ tests/queue/apis/queue.test.ts
@@ -183,6 +183,20 @@ describe("Queue APIs test", () => {
           startsOn: new Date("2017-12-31T11:22:33.4567890Z")
         },
         id: "policy2"
+      },
+      {
+        accessPolicy: {
+          permissions: "ra",
+          startsOn: new Date("2017-12-31T11:22:33.4567890Z")
+        },
+        id: "policy3"
+      },
+      {
+        accessPolicy: {
+          expiresOn: new Date("2030-11-31T11:22:33.4567890Z"),
+          permissions: "up"
+        },
+        id: "policy4"
       }
     ];